修改积分数据类型以及更新新用户注册积分

+++ ark-credit-service/src/main/java/com/cryptal/ark/arkcreditservice/config/service/ConfigService.java
@@ -7,11 +7,11 @@ public interface ConfigService {
      * 邀请新用户奖励
      * @return
      */
-    int getInviteUserReward();
+    Long getInviteUserReward();
 
     /**
      * 新注册用户奖励
      * @return
      */
-    int getNewUserReward();
+    Long getNewUserReward();
 }
